+++ ml/tests/unit/training/test_xgboost_trainer_unit.py
@@ -87,9 +87,9 @@ class TestXGBoostTrainerModelTraining:
 
         rng = np.random.default_rng(42)
         X_train = rng.standard_normal((80, 2))
-        y_train = rng.integers(0, 2, 80)
+        y_train = rng.integers(0, 2, 80).astype(np.float64)
         X_val = rng.standard_normal((20, 2))
-        y_val = rng.integers(0, 2, 20)
+        y_val = rng.integers(0, 2, 20).astype(np.float64)
 
         # Mock XGBoost components
         mock_dtrain = MagicMock()
@@ -126,9 +126,9 @@ class TestXGBoostTrainerModelTraining:
 
         rng = np.random.default_rng(42)
         X_train = rng.standard_normal((80, 2))
-        y_train = rng.integers(0, 2, 80)
+        y_train = rng.integers(0, 2, 80).astype(np.float64)
         X_val = rng.standard_normal((20, 2))
-        y_val = rng.integers(0, 2, 20)
+        y_val = rng.integers(0, 2, 20).astype(np.float64)
 
         # Mock XGBoost components
         mock_xgb.DMatrix.return_value = MagicMock()
